requirements.txt
aws-cdk-lib==2.12.0
constructs>=10.0.0,<11.0.0
boto3
regex
requests
BeautifulSoup4